Changed Long to Integer to match with Model

+++ src/main/java/springboot/Model/GroupRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "groups", path = "groups")
-public interface GroupRepository extends CrudRepository<Group, Long> {
+public interface GroupRepository extends CrudRepository<Group, Integer> {
 
 }
